fix color array

+++ native_raycasting/include/texture.h
@@ -1,10 +1,10 @@
 #pragma once
 #include <dmsdk/sdk.h>
 struct Color {
-	int r;
-	int g;
-	int b;
-	int a;
+	uint8_t r;
+	uint8_t g;
+	uint8_t b;
+	uint8_t a;
 };
 
 struct Texture {
